requirements.txt
flask
flask_restful
mysql-connector-python
python-dotenv